webkit/data/layout_test_results/kjs/LayoutTests/plugins/embed-attributes-setting-expected.txt

[Embed is element specified in markup]
PASS: embed.getAttribute('align') should be 1 and is.
PASS: embed.getAttribute('height') should be 1 and is.
PASS: embed.getAttribute('name') should be 1 and is.
PASS: embed.getAttribute('width') should be 1 and is.
PASS: embed.getAttribute('type') should be 1 and is.
PASS: embed.getAttribute('src') should be 1 and is.
FAIL: typeof embed.Play should be function but instead is undefined.
----------
[Embed is dynamically created element with only type specified]
FAIL: typeof embed.Play should be function but instead is undefined.
----------
[Embed is dynamically created element with only src specified]
FAIL: typeof embed.Play should be function but instead is undefined.
 
